+++ examples/mlp_random_tensorflow2/server.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-# Copyright (c) 2022, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2022-2023, NVIDIA CORPORATION. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -57,6 +57,7 @@ with Triton() as triton:
             Tensor(name="output", dtype=np.float32, shape=(224, 224, 3)),
         ],
         config=ModelConfig(max_batch_size=16),
+        strict=True,
     )
     logger.info("Serving inference")
     triton.serve()