+++ pytriton/models/model.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2022, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2022-2023, NVIDIA CORPORATION. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -75,6 +75,7 @@ class Model:
         config: ModelConfig,
         workspace: Workspace,
         triton_context: TritonContext,
+        strict: bool,
     ):
         """Create Python model with required data.
 
@@ -86,6 +87,7 @@ class Model:
             outputs: Model outputs definition
             config: model configuration parameters
             workspace: workspace for storing artifacts
+            strict: Enable strict validation of model outputs
 
         Raises:
             PyTritonValidationError if one or more of provided values are incorrect.
@@ -97,6 +99,7 @@ class Model:
         self.zmq_context = zmq.Context()
         self._observers_lock = threading.Lock()
         self._inference_handlers_lock = threading.Lock()
+        self._strict = strict
 
         self.infer_functions = [inference_fn] if isinstance(inference_fn, Callable) else inference_fn
         if not isinstance(self.infer_functions, (Sequence, Callable)):
@@ -162,6 +165,7 @@ class Model:
                         model_config=triton_model_config,
                         shared_memory_socket=f"{self._shared_memory_socket}_{i}",
                         zmq_context=self.zmq_context,
+                        strict=self._strict,
                     )
                     inference_handler.on_proxy_backend_event(self._on_proxy_backend_event)
                     inference_handler.start()